+++ test/fixtures/fake_all_observer/lib/all_observer.dart
@@ -16,6 +16,8 @@ class CoreObservable<T> {
 
 class Observable<T> extends CoreObservable<T> {
   Observable(super.value);
+
+  static void batch(void Function() callback) => callback();
 }
 
 class ObservableList<T> extends CoreObservable<List<T>> {
@@ -55,21 +57,25 @@ Disposer effect(void Function() callback) => Disposer();
 Disposer ever<T>(
   CoreObservable<T> observable,
   void Function(T value) onChange,
-) => Disposer();
+) =>
+    Disposer();
 Disposer once<T>(
   CoreObservable<T> observable,
   void Function(T value) onChange,
-) => Disposer();
+) =>
+    Disposer();
 Disposer debounce<T>(
   CoreObservable<T> observable,
   void Function(T value) onChange, {
   Duration time = const Duration(milliseconds: 300),
-}) => Disposer();
+}) =>
+    Disposer();
 Disposer interval<T>(
   CoreObservable<T> observable,
   void Function(T value) onChange, {
   Duration time = const Duration(seconds: 1),
-}) => Disposer();
+}) =>
+    Disposer();
 
 void batch(void Function() callback) => callback();